Move chip to widget subfolder
